Added jwt context with a working example
