+++ thruster/Cargo.toml
@@ -102,3 +102,4 @@ smallvec = "0.6.2"
 snafu = "0.4.1"
 prost = "0.5.0"
 http = "0.2"
+frank_jwt = "3.1.2"
